requirements.txt
black==22.3.0
mysql-connector-python==8.0.29
numpy==1.22.3
pandas==1.4.2
pyflakes==2.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
